move data property into the base EnvironmentForm so that NO_ACCESS is converted to None when creating a new application member
