+++ atst/forms/data.py
@@ -113,7 +113,7 @@ TEAM_EXPERIENCE = [
 
 ENV_ROLE_NO_ACCESS = "No Access"
 ENV_ROLES = [(role.value, role.value) for role in CSPRole] + [
-    (ENV_ROLE_NO_ACCESS, "No access")
+    (ENV_ROLE_NO_ACCESS, ENV_ROLE_NO_ACCESS)
 ]
 
 JEDI_CLIN_TYPES = [
