+++ atst/forms/team.py
@@ -3,7 +3,7 @@ from wtforms.fields import FormField, FieldList, HiddenField, RadioField, String
 from wtforms.validators import Required
 
 from .application_member import EnvironmentForm as BaseEnvironmentForm
-from .data import ENV_ROLES, ENV_ROLE_NO_ACCESS as NO_ACCESS
+from .data import ENV_ROLES
 from .forms import BaseForm
 from atst.forms.fields import SelectField
 from atst.domain.permission_sets import PermissionSets
@@ -17,13 +17,6 @@ class EnvironmentForm(BaseEnvironmentForm):
         default=None,
         filters=[lambda x: None if x == "None" else x],
     )
-
-    @property
-    def data(self):
-        _data = super().data
-        if "role" in _data and _data["role"] == NO_ACCESS:
-            _data["role"] = None
-        return _data
 
 
 class PermissionsForm(FlaskForm):
